feat: implement mobile first design
