+++ src/routes/home.tsx
@@ -1,8 +1,10 @@
-import Header from "../components/Header";
+import MobileFooter from "@/components/MobileFooter";
+import MobileHeader from "@/components/MobileHeader";
 import Summary from "../components/Summary";
 import TransactionsTable from "../components/TransactionsTable";
 import '../globals.css'
 import { Toaster } from "sonner";
+import Header from "@/components/Header";
 
 export default function Home() {
     return (
@@ -11,9 +13,13 @@ export default function Home() {
 
             <Header />
 
+            <MobileHeader />
+
             <Summary />
 
             <TransactionsTable />
+
+            <MobileFooter />
         </>
     )
 }
